Change u64 to u32

+++ model/src/channel/mod.rs
@@ -117,7 +117,7 @@ pub struct Channel {
     pub member_count: Option<u8>,
     /// Number of messages in the channel.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub message_count: Option<u64>,
+    pub message_count: Option<u32>,
     /// Name of the channel.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub name: Option<String>,
@@ -367,7 +367,7 @@ mod tests {
                 user_id: Some(Id::new(5)),
             }),
             member_count: Some(50_u8),
-            message_count: Some(50_u64),
+            message_count: Some(50),
             name: Some("newsthread".into()),
             newly_created: Some(true),
             nsfw: None,
@@ -448,7 +448,7 @@ mod tests {
                 user_id: Some(Id::new(5)),
             }),
             member_count: Some(50_u8),
-            message_count: Some(50_u64),
+            message_count: Some(50),
             name: Some("publicthread".into()),
             newly_created: Some(true),
             nsfw: None,
@@ -530,7 +530,7 @@ mod tests {
                 user_id: Some(Id::new(5)),
             }),
             member_count: Some(50_u8),
-            message_count: Some(50_u64),
+            message_count: Some(50),
             name: Some("privatethread".into()),
             newly_created: Some(true),
             nsfw: None,
